requirements.txt
Django==1.7
django-appconf==0.6
django-compressor==1.4
factory-boy==2.4.1
pyScss==1.2.1
pylibmc==1.3.0
pytz==2014.7
six==1.8.0
wsgiref==0.1.2
